Add `image.source`, use it for `from_sendstream`
Summary:
The idea behind `image.source` is that throughout our build toolchain, we want to refer to files or directories in the following cases:
 - a file or directory that **is** the output of a vanilla Buck target
 - a file or directory that is **contained** in the directory output by a vanilla Buck target
 - a file or directory that is contained in a btrfs subvolume output by an `image.layer` Buck target

From the point of view of the function consuming the file or directory, these situations should be largely identical, so the abstraction here aims to separate the concern of specifying the "what to act on" from the "how to act on it".

NB: I say "largely identical" because vanilla Buck outputs do not have very meaningful `stat` metadata. I think one can at best rely on the 3 bits of "is {readable,writable,executable} by repo user?", with everything else being non-reproducible as far as Buck is concerned. In contrast, the content of subvolumes should have deterministic `stat` metadata (except for timestamps, which we do not yet normalize).

Reviewed By: zeroxoneb

Differential Revision: D16939493

fbshipit-source-id: 239d7a843a14bcfea7b88a06d53c35c3093b1ffb