Fix @mode/opt build for the `find-built-subvol` binary.
Summary:
Fix the mode/opt build for the `find-build-subvol` binary.  With the
addition of the `image.source` macro and it's use for building layers from
sendstreams we now have a dependency on the `find-built-subvol` binary during
build evaluation/execution.  This means we have to use a `zip` par instead of
`fastzip` because we now rely on the `set_up_volume.sh` script transitively for
building layers.

Differential Revision: D16965174

fbshipit-source-id: afaa57afec2044d7782acda5e41def75c882fe1f