`image.source` support for `image.feature`
Summary:
D16939493 introduced `image.source`, so read that for the rationale. Here, we make the abstraction meaningfully useful by updating the more frequently used `image.feature`s to support it:
  - `install_data`
  - `install_executables`
  - `rpms`

One practical implication of this is that we'll be able to use `image.rpmbuild_layer` to produce some RPMs, and then to install them straight from that layer using the `rpms` action.

This omits `image.source` support from `tarballs` until desired. But that change would be easy.

Reviewed By: anitazha

Differential Revision: D16939492

fbshipit-source-id: 68bad4bbb3b1b47164d72be4217315731813b489