Use slimos stable in fs_image tests to speed things up.
Summary:
Use the slimos-stable image in test to speed things up.

But also, copy systemd exec wrapper with `--dereference`.  Due to how the
`artifacts_may_require_repo` behavior works when we use an `image.layer` that
is built using `from_sendstream` we won't get artifacts bind mounted into the
container.  in the `mode/dev` case, the `wrap_systemd_exec.sh` script is
actually a symlink inside the `nspawn_in_subvol` par, which doesn't work unless
the repo is bind mounted int.  Using `--dereference` on the `cp` makes this not
matter and I think is relatively safe.

Also, also due to how the `artifacts_may_require_repo` behavior works we need to
know when we run a test if we should bind the repo and so we add this to the unit test
helper.

Reviewed By: filbranden, snarkmaster

Differential Revision: D16933973

fbshipit-source-id: e386fb88438818c18fa8416ecd47f49e14e044e5