Replace `image.layer(from_sendstream)` with `image.sendstream_layer`
Summary:
We are about to get more `layer` types, so the already-smelly API of having "mutually exclusive groups of kwargs" would become completely unusable.

Moreover, having a separate entry point just overall results in cleaner code.

This is intended to be a no-op refactor. The only change in function is that the build log headers become a bit more descriptive.

Reviewed By: zeroxoneb

Differential Revision: D16996952

fbshipit-source-id: 6bc0518489511fb155596f97bfe0b60af4ed179d