Make it possible to add other item types that create the initial subvol
Summary:
This is needed to add a `ReceiveSendstreamItem`.

Before this refactor, `FilesystemRootItem` and `ParentLayerItem` have TWO distinct jobs — (i) creating the subvolume, (ii) reporting what that subvolume "provides" to the dependency-ordering step.

Here, I split the "provides" part into `PhasesProvideItem`, which I think makes the code and flow significantly clearer.

It also makes it so that the `MAKE_SUBVOL`-phase items are no longer treated specially (the only tiny exception is that `FilesystemRootItem` is defaulted in `dep_graph.py`).

As a result, adding `ReceiveSendstreamItem` will be easy.

Reviewed By: zeroxoneb

Differential Revision: D17098331

fbshipit-source-id: 8624ac5e81b3fb7690b166b56274888605b692c7