New sigil for uniformly resolving layer targets in compiler inputs
Summary:
There are a few places in the compiler code where we'd `find_built_subvol` inside an `ImageItem` implementation, based on a layer target path passed from above.

This was OK, but rather annoying to test, because to construct such an `ImageItem`, one has to interact with the `SubvolumeOnDisk` serialization layer, which does strict validation that makes it hard to use layers that are not actual Buck "layer" target outputs.

In the new world, `ImageItem`s can actually **contain** a `Subvol`. This is easy to test via `TempSubvolumes`.

One downside of items containing a non-POD object like `Subvol` is that this can break implicit item deduplication until we add `Subvol.__eq__` and `Subvol.__hash__`. I am not doing so in this diff because I don't actually know of a case where we should care about this. Once it occurs, here is a possible fix:

```
    # Adding hash & comparison since `Subvol` can be contained in `ImageItem`s
    # Without this, two semantically identical `Subvol`s would not compare
    # equal, and item deduplication would not take place.

    def __eq__(self, other):
        if not isinstance(other, Subvol):
            return NotImplemented
        if self._path != other._path:
            return False
        assert self._exists == other._exists
        return True

    def __hash__(self):
        return hash(self._path)
```

Another possible fix is to make `Subvol` a `NamedTuple` containing just `path`, and use a look-aside dictionary to access the mutable state of `exists`. Not discussing the tradeoffs here since I'm not sure if this will come up.

Reviewed By: zeroxoneb

Differential Revision: D17098330

fbshipit-source-id: 495c39fe740886444af59eed2587e512ce1cae5f