Allow changing subvol name in demo sendstreams
Summary:
This is a test-only change.

After some deliberation, I decided that `ReceiveSendstreamItem` should conform to the existing `image.layer` contract of being able to choose the resulting subvolume name.

In tests, we'll be receiving something like `create_ops`, but then overriding the sendstream-supplied subvol-name to something else.

To validate the resulting subvol, I need the demo rendering to accept non-built-in names for the subvolumes.

Reviewed By: zeroxoneb

Differential Revision: D17098328

fbshipit-source-id: 25917be1afbf788295f2b6c5a456bbb5800409d5