Allow layer mount configs to be generated
Summary:
To support packaged layers, the mount config has to come from the output of another Buck target.

Update: in my latest factoring of `fetched_package_layer` this is not a literal requirement, but doing it that way still makes for a better separation of responsibilities in the code, so I'm keeping this diff.

Reviewed By: zeroxoneb

Differential Revision: D17353277

fbshipit-source-id: 97b0b2811f0b4989c53a10f1ab41a5f9709f15c1