Fix ZSTD tarball extraction
Summary:
The old code here was incorrect in that we would try to parse the TAR stream coming in from the `zstd` subprocess **after** exiting the `zstd` process's context. The exit, per the docs, would "close standard file descriptors, and wait for the process."

So we would essentially end up closing the input stream to `tarfile.open`, leading to:

```
  File "/usr/lib/python3.6/tarfile.py", line 569, in __read
    buf = self.fileobj.read(self.bufsize)
ValueError: read of closed file
```

Unit tests didn't catch this because it requires the uncompressed stream to be larger than the intermediate buffers. I'm not inclined to add a test with a bigger ZST tarball simply because it's a game of guessing buffer sizes on the test systems, and a large enough tarball will make the test heavy. Reviewers, feel free to convince me otherwise.

Reviewed By: williampsmith

Differential Revision: D17772948

fbshipit-source-id: a60e91a5936a918ba8c81d16806953b9e3272ba0