Unified function for transparent decompression based on extension
Summary:
This adds support for receiving compressed sendstreams.

We now use the same implementation for tarball unpacking and for sendstream unpacking.

In effect, this also drops support for some of the formats that both GNU `tar` and Python `tarfile` would previously autodetect (`bz2` and `lzma`), but this is kind of a moot point because `zstd` dominates both of those in practice.

Note that some of the usage would get easier if we decompressed based on magic headers instead of extension, but implementing that seems like more hassle, which isn't worthwhile for now.

Reviewed By: zeroxoneb

Differential Revision: D17800414

fbshipit-source-id: ab117b18a0fe7fcf6f17f14a07235417264b1293