Improve language for Buck building container filesystems
Summary:
Use functions with named arguments rather than tuples, lists and hashes with special key names. (This is similar to the proposal named "factory functions" in T46837515.)

Use singular rather than plural almost everywhere (with the one exception for {,un}install_rpms which takes a list of packages.) This makes it simpler to reason about multiple separate actions, allows reordering related groups together and makes it possible to really streamline the API. It is still possible to perform the same action on multiple items by using Python generator expressions to apply the same action to a list of elements.

Flatten the hierarchy (or, perhaps more precisely, do not deepen it.) In particular, "mode", "user" and "group" are still attributes of the actions rather than a separate stat objects. Split attributes that are independent and could be specified on their own, in particular: "user" vs. "group".

Use boolean named attributes rather than magic strings wherever possible (e.g. "if_exists" for remove.) This could potentially allow for better linting or syntax checking. Unfortunately Buck's Python dialect doesn't allow for keyword-only arguments or full type checking support, which would be helpful here, but the intention is to have those used only as named arguments.

Update all consumers of this language to use the new constructs.

The underlying model still reflects the old language. This is intentional, both in order to allow discussion on this commit to move faster, to avoid scope creep and to reduce risk in introducing this change. The intention is to land this commit without touching the underlying model and later refactor it to simplify it to support the new language.

 ---

Updates since initial version:

- Made `parent` argument to `image.mkdir()` mandatory and moved it to the first argument, so the path is written in the order it's used.

- Inverted the order of arguments to (source, dest) in `install_data`, `install_executable` and `tarball`, so it matches the normal expected interface.

- Replaced `image.mount()` with `image.layer_mount()` and made `image.host_{file,dir}_mount()` return an image feature, so they can be used standalone.

- Replaced the optional arguments in `image.remove()` with a single `must_exist = True` which now needs an override to False for optional removals.

- Otherwise, fixed some minor nits.

Reviewed By: snarkmaster

Differential Revision: D17772674

fbshipit-source-id: 26ddf4d615fec7801df070d4eb0344d934e2e8a0