Automatically export_file when needed.
Summary:
This is a simplification that has the Buck image build system run export_file() for us whenever it's actually needed.

Detection is done based on whether the source looks like a Buck target (in specific, whether it has a ":" in the name) or like a filename. If the latter, then generate a Buck target name and call export_file() for it.

For cases not directly supported by this (e.g. filename contains a colon, need to access the file externally, so need a specific visibility, or need to include the same file twice), let's just keep using normal export_file() rules.

Reviewed By: zeroxoneb, snarkmaster

Differential Revision: D17944589

fbshipit-source-id: 9e63c310945bafdf502bde173302ebc30c86740b