Allow duplicate uses of maybe_export_file()
Summary:
One limitation of `maybe_export_file()` is that it can be used only once for a certain path, the second use will create a duplicate Buck target that will cause Buck to error out and break the build.

This has been inconvenient in a number of situations, with one pre-existing the first push for `maybe_export_file()` and two more appearing since the commit landed last week.

Let's check if a rule already exists (which we assume was generated by a previous `maybe_export_file()` and reuse it if it's already there.

Use a sigil prefix `_IMAGE_EXPORT_FILE__` for the Buck target name, to avoid clashes with targets defined by the user.

Reviewed By: zeroxoneb

Differential Revision: D18172230

fbshipit-source-id: 3f585fef62ce3c5436a6b688c4b7ef11c869eeaa