Fix handling /var/cache/yum when installing rpm by build_appliance
Summary:
Replacing /mnt with /work in D15492258, I forgot to update "mkdir -p" and "mount --bind" parts of systemd_nspawn args.
This diff fixes relevant piece of RpmActionItem and updates the unit test testing build_appliance: our expectation is that /var/cache/yum must not be polluted by RpmActionItem, hence, in this test evnironment it is empty before and after applying RpmActionItem, and rmdir must be enough to remove it. If RpmActionItem changes this behavor in future, the test would start failing because rmdir fails if applied to non-empty dir.

Reviewed By: filbranden, snarkmaster

Differential Revision: D18393133

fbshipit-source-id: e7122d06f947322e470c09ee11a702748b0ae3b6