Stop clobbering /etc/machine-id and leave it as is
Summary:
Our container builds should ship an empty `/etc/machine-id` file, since that's the recommendation for container builds.

Not shipping a `machine-id` file at all triggers "firstboot" behavior, which is not limited to running `systemd-firstboot.service` (which we can override) but also to running `systemctl preset-all` from the system manager itself during boot (which is hardcoded.)

The man page for machine-id itself recommends that container images should simply ship an `/etc/machine-id` file.

Remove the code that was trying to enforce obfuscation by bind-mounting `/dev/null` on top of it. The file will typically be empty, so there's no harm in removing the safety net.

The tests that check for existence/contents of the machine-id file have been updated. I removed the tests that are checking that on host mount and I'm checking that on slimos specifically.

Reviewed By: zeroxoneb

Differential Revision: D18098816

fbshipit-source-id: 57ee9a169879ecb4990a38e046ed7f7591052eb1