Small refactor to build-appliance path of RpmActionItem
Summary:
In the near future, all RPM installation will go through the build appliance code path. To make this happen, we have to be able to use a build appliance from yum/dnf tests. This change uncouples the compiler-specific logic from the actual action of running a yum/dnf command inside a build appliance.

There is a small change in function — instead of using `/work` for the workdir, I am now using `/work<UUID>` to make sure we don't collide with anything, and to ensure we don't hardcode `/work` in weird places.

Differential Revision: D18612123

fbshipit-source-id: 2defe3b7b79d664c0782ca54c0c4129aeb1dbb98