Convert from py2->bzl
Summary:
I believe the only reason this is still marked as `.py` with `python2`
as an interperter is to allow it to behave as part of the lib introduced in
D3520155 which makes the target determinator behave correctly for generated
TARGETS files. Or at least this was the limitation that it overcame at the
time. Is this TD hack still necessary? Anyway, this cleans up its "python-ness"
by moving to a .bzl suffix, packaging as a resource to avoid the `.py`
complaint from `python_library`. It also replaces
`set` with `skylib:sets` and replaces the `constants.py`
include_defs with a quine'd version in `constants.bzl`.

The two things I wasn't able to fix are:
1) Usage of functools (`skylib:partial` doesn't return a functor, and skylark
doesn't have lambda support, so there's no good replacement for this)
2) Usage of `globals()` when installing symbols

Reviewed By: andrewjcg

Differential Revision: D21540240

fbshipit-source-id: 3109e4e6f2541338271bf42ca79496f7d536ef3f