ide script template
Summary:
## Context

I introduced the HaskellIdeRule last week as a fork of the Ghci rule in D25461671

## This stack

The goal is to be able to create haskell_ide targets in Starlark that produce a compilation database to load *multiple* Haskell projects from source in ghcide

## This diff

Adds a new script template for use in HaskellIdeRule instead of reusing the ghci template. The template is put in use by the next Diff in this stack.

Instead of invoking ghci, this template simply prints the set of flags and targets, one per line, to stdout

Reviewed By: myszon

Differential Revision: D25723154

fbshipit-source-id: d031792c29081526b19b5c0b3771ae2e9a3106fa